Extract remote data source interface
