+++ app/src/main/java/com/fjoglar/lyricly/data/SongsRepository.java
@@ -19,7 +19,6 @@ package com.fjoglar.lyricly.data;
 import androidx.annotation.Nullable;
 
 import com.fjoglar.lyricly.data.model.Song;
-import com.fjoglar.lyricly.data.source.remote.entity.Track;
 
 import java.util.Date;
 import java.util.List;
@@ -27,59 +26,40 @@ import java.util.List;
 import io.reactivex.Completable;
 import io.reactivex.Flowable;
 
-public class SongsRepository implements SongsDataSource.LocalDataSource,
-        SongsDataSource.RemoteDataSource, SongsDataSource.PreferencesDataSource {
+public class SongsRepository implements SongsDataSource.LocalDataSource, SongsDataSource.PreferencesDataSource {
 
     @Nullable
     private static SongsRepository INSTANCE = null;
 
     private final SongsDataSource.LocalDataSource mSongsLocalDataSource;
-    private final SongsDataSource.RemoteDataSource mSongsRemoteDataSource;
     private final SongsDataSource.PreferencesDataSource mPreferencesDataSource;
 
     private SongsRepository(SongsDataSource.LocalDataSource songsLocalDataSource,
-                            SongsDataSource.RemoteDataSource songsRemoteDataSource,
                             SongsDataSource.PreferencesDataSource preferencesDataSource) {
         mSongsLocalDataSource = songsLocalDataSource;
-        mSongsRemoteDataSource = songsRemoteDataSource;
         mPreferencesDataSource = preferencesDataSource;
     }
 
     /**
      * Returns the single instance of this class, creating it if necessary.
      *
-     * @param songsLocalDataSource  the backend data source
-     * @param songsRemoteDataSource the device storage data source
+     * @param songsLocalDataSource the backend data source
      * @return the {@link SongsRepository} instance
      */
     public static SongsRepository getInstance(SongsDataSource.LocalDataSource songsLocalDataSource,
-                                              SongsDataSource.RemoteDataSource songsRemoteDataSource,
                                               SongsDataSource.PreferencesDataSource preferencesDataSource) {
         if (INSTANCE == null) {
-            INSTANCE = new SongsRepository(songsLocalDataSource,
-                    songsRemoteDataSource,
-                    preferencesDataSource);
+            INSTANCE = new SongsRepository(songsLocalDataSource, preferencesDataSource);
         }
         return INSTANCE;
     }
 
     /**
-     * Used to force {@link  #getInstance(SongsDataSource.LocalDataSource,
-     * SongsDataSource.RemoteDataSource, SongsDataSource.PreferencesDataSource)} to create a new
-     * instance next time it's called.
+     * Used to force {@link  #getInstance(SongsDataSource.LocalDataSource, SongsDataSource.PreferencesDataSource)}
+     * to create a new instance next time it's called.
      */
     public static void destroyInstance() {
         INSTANCE = null;
-    }
-
-    @Override
-    public List<Track> fetchTopSongs(int limit) {
-        return mSongsRemoteDataSource.fetchTopSongs(limit);
-    }
-
-    @Override
-    public String fetchSongLyrics(String artist, String title) {
-        return mSongsRemoteDataSource.fetchSongLyrics(artist, title);
     }
 
     @Override
