+++ app/src/main/java/com/fjoglar/lyricly/songs/top/TopSongsViewModel.java
@@ -21,6 +21,7 @@ import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
 import com.fjoglar.lyricly.data.SongsRepository;
+import com.fjoglar.lyricly.data.source.remote.SongsRemoteDataSource;
 import com.fjoglar.lyricly.songs.SongsResponse;
 import com.fjoglar.lyricly.songs.SongsViewModel;
 import com.fjoglar.lyricly.util.schedulers.SchedulerProvider;
@@ -30,6 +31,7 @@ import io.reactivex.disposables.CompositeDisposable;
 public class TopSongsViewModel extends ViewModel implements SongsViewModel {
 
     private SongsRepository mSongsRepository;
+    private SongsRemoteDataSource mSongsRemoteDataSource;
 
     private final CompositeDisposable mDisposables = new CompositeDisposable();
 
@@ -37,8 +39,9 @@ public class TopSongsViewModel extends ViewModel implements SongsViewModel {
     private final MutableLiveData<Boolean> mLoadingState = new MutableLiveData<>();
 
 
-    public TopSongsViewModel(SongsRepository songsRepository) {
+    public TopSongsViewModel(SongsRepository songsRepository, SongsRemoteDataSource songsRemoteDataSource) {
         mSongsRepository = songsRepository;
+        mSongsRemoteDataSource = songsRemoteDataSource;
         updateTopSongs();
         getTopSongs();
     }
@@ -59,7 +62,7 @@ public class TopSongsViewModel extends ViewModel implements SongsViewModel {
     }
 
     private void updateTopSongs() {
-        mDisposables.add(new UpdateTopSongsUseCase().execute(mSongsRepository, null)
+        mDisposables.add(new UpdateTopSongsUseCase().execute(mSongsRepository, mSongsRemoteDataSource)
                 .subscribeOn(SchedulerProvider.getInstance().io())
                 .observeOn(SchedulerProvider.getInstance().ui())
                 .doOnSubscribe(__ -> mLoadingState.setValue(true))
