+++ app/src/main/java/com/fjoglar/lyricly/songs/top/UpdateTopSongsUseCase.java
@@ -19,6 +19,7 @@ package com.fjoglar.lyricly.songs.top;
 import com.fjoglar.lyricly.data.SongsRepository;
 import com.fjoglar.lyricly.data.model.Song;
 import com.fjoglar.lyricly.data.source.mapper.SongDataMapper;
+import com.fjoglar.lyricly.data.source.remote.SongsRemoteDataSource;
 import com.fjoglar.lyricly.data.source.remote.entity.Track;
 import com.fjoglar.lyricly.util.usecases.CompletableUseCase;
 
@@ -27,19 +28,19 @@ import java.util.List;
 
 import io.reactivex.Completable;
 
-public class UpdateTopSongsUseCase implements CompletableUseCase<Void> {
+public class UpdateTopSongsUseCase implements CompletableUseCase<SongsRemoteDataSource> {
 
     private long DAY_IN_MILLIS = 24 * 60 * 60 * 1000;
 
     @Override
-    public Completable execute(SongsRepository repository, Void parameter) {
+    public Completable execute(SongsRepository repository, SongsRemoteDataSource remoteDataSource) {
         return Completable.fromAction(() -> {
 
             if (System.currentTimeMillis() - repository.getLastUpdatedTimeInMillis() < DAY_IN_MILLIS) {
                 return;
             }
 
-            List<Track> tracks = repository.fetchTopSongs(200);
+            List<Track> tracks = remoteDataSource.fetchTopSongs(200);
             if (tracks.isEmpty()) {
                 return;
             }
@@ -50,7 +51,7 @@ public class UpdateTopSongsUseCase implements CompletableUseCase<Void> {
                 Song song = repository.getTopSongByNapsterId(track.getId());
 
                 if (song == null) {
-                    String lyrics = repository.fetchSongLyrics(track.getArtistName(), track.getName());
+                    String lyrics = remoteDataSource.fetchSongLyrics(track.getArtistName(), track.getName());
                     if (lyrics != null && !lyrics.isEmpty()) {
                         repository.saveSong(SongDataMapper.transform(track, tracks.indexOf(track), lyrics));
                     }
